NannysLM/backend/requirements.txt
# Backend NannysLM - FastAPI
# Dependencias necesarias para el proyecto

# Framework principal
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de datos
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.13.1

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validación y serialización
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS y middleware
python-cors==1.7.0

# Utilidades
python-dotenv==1.0.0
email-validator==2.1.0

# Desarrollo y testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2